+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog for aim.models
 ------------------
 
 - Added default to IS3Bucket.policy
+- Added Route53 to schema and model
 
 
 1.0.1 (2019-06-19)